net/http: improve documentation of a Client request
The documentation of Client.Do is clear, that the it is not enough
to close the responce body. It als has to be read.

The existing first description of this case only said, that it has
to be closed.

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

Change-Id: I077fb84b1adf9581a59647075b222655301ff07c
GitHub-Last-Rev: ecbfa08ab49de170fcf127304509b7f9cf43dfb0
GitHub-Pull-Request: golang/go#50515