runtime: eliminate arbitrary timeout in TestStackGrowth
Instead, allow the test to run up until nearly the test's deadline,
whatever that may be, and then crash with a panic (instead of calling
t.Errorf) to get a useful goroutine dump.

With the arbitrary timeout removed, we can now also run this test in
short mode, reducing its impact on test latency.

Also pad the object on which we set the finalizer to be large enough
to not use the tiny allocator (compare #48968), to ensure that the
finalizer will eventually run even on 32-bit systems.

Fixes #19381

Change-Id: Ie1fae321a2973fcb9b69a012103363f16214f529