[DO-NOT-REVIEW] for test
cmd/internal/obj/arm64: remove useless unsafe point marking

When the frame size is too large for a MOVD.W instruction, in the
prologue of the function, we use the following three instructions to
save link register (R30) and adjust SP.
    SUB     $OFFSET, RSP, R27
    MOVD    R30, (R27)
    MOVD    R27, RSP
The sequence is not async preemtible because the saved LR may be
clobbered, so we mark this by explicitly inserting PCDATA instructoins
before and after the sequence. However this is not necessary because
every instruction in this sequence uses the R27 register, which causes
the instruction that uses it to be marked as unsafe point. So this CL
removes these two unnecessary two function calls.

Change-Id: I6efaedf92b62f356dc8962b350bb7fe6d7eb30cc