reflect: use binary search in MethodByName
The only case where a binary search loses to a linear search
is when the search item is somewhere in the beginning of the slice.
In general, that's not an issue as the searched method can't
be expected to be in some specific position. But it still helps
to have a separate path with linear search to avoid the slowdown
for small slices (which are not uncommon).

This linear search is a big problem for types with large number of methods.
I found that using binary search helps hugo static site builder to execute
faster (16.5s vs 20.4s).

See the referenced issue for more details and stats.

	name                           old time/op  new time/op  delta
	MethodByName/4_first-8          426ns ± 2%   423ns ± 1%     ~     (p=0.056 n=10+9)
	MethodByName/4_last-8           482ns ± 1%   476ns ± 1%   -1.33%  (p=0.000 n=10+10)
	MethodByName/4_nonexisting-8   57.8ns ± 0%  57.3ns ± 1%   -0.78%  (p=0.000 n=9+9)
	MethodByName/16_first-8         427ns ± 1%   538ns ± 0%  +25.97%  (p=0.000 n=9+9)
	MethodByName/16_last-8          643ns ± 1%   519ns ± 1%  -19.24%  (p=0.000 n=9+9)
	MethodByName/16_nonexisting-8   194ns ± 0%   105ns ± 0%  -46.03%  (p=0.000 n=8+9)
	MethodByName/32_first-8         429ns ± 1%   552ns ± 1%  +28.78%  (p=0.000 n=10+8)
	MethodByName/32_last-8          831ns ± 0%   540ns ± 2%  -34.97%  (p=0.000 n=10+9)
	MethodByName/32_nonexisting-8   396ns ± 1%   125ns ± 1%  -68.56%  (p=0.000 n=10+10)
	MethodByName/64_first-8         431ns ± 1%   580ns ± 2%  +34.64%  (p=0.000 n=10+10)
	MethodByName/64_last-8         1.36µs ± 1%  0.56µs ± 1%  -58.95%  (p=0.000 n=10+8)
	MethodByName/64_nonexisting-8   759ns ± 0%   146ns ± 1%  -80.82%  (p=0.000 n=8+10)
	[Geo mean]                      430ns        303ns       -29.58%

Refs #50617

Change-Id: I5d47459e78acdeaf92b09dae0a22f5df1973eccc