cmd/compile/internal/types2: report an error when using a broken alias
The type checker doesn't have a general mechanism to "use" the type
of a type alias whose type depends on a recursive type declaration.
In some cases, the type of a type alias is needed before it is
determined; the type is invalid in that case but no error is reported.
This can lead to subsequent crashes in the compiler.

An always-correct fix will likely require some form of forwarding
type which is a fairly pervasive change and also may affect the
type checker API.

This CL introduces a simple side table, a map of broken type aliases
that is consulted before the type associated with a type alias is
used. If the type alias is broken, an error is reported.
This is a stop-gap solution that prevents the compiler from crashing.
The reported error refers to the corresponding issue which suggests
a work-around that may be applicable in some cases.

Also fix a minor error related to type cycles: If we have a cycle
that doesn't start with a type, don't use a compiler error message
that explicitly mentions "type".

Fixes #50259.
Fixes #50276.
For #50729.

Change-Id: Ie8e38f49ef724e742e8e78625e6d4f3d4014a52c