cmd/go: evaluate root symlink in matchPackages
This change is a roll forward of CL 380057, which was reverted in
380294, because it was breaking an x/tools test that used overlays on
top of a directory path that contained a symlink. The underlying test
has been fixed in this rollforward, and the test case for the CL has
been expanded with an overlay to test against regressions of that case.

Original change description:

This fixes checks for crossing module boundaries when the root of
the module is a symlink. We're comparing paths by string, so we need
to follow the symlink to get the proper path to compare.

Change-Id: I30715a03a66fa44a8927b4385231dc7ff9d4fe0b