debug/dwarf: fix problems with handling of bit offsets for bitfields
DO NOT SUBMIT

[Needs a better set of regression tests, better comment].

This patch reworks the handling of the DWARF DW_AT_bit_offset and
DW_AT_data_bit_offset attributes to resolve problems arising from
a previous related change (CL 328709).

In CL 328709 the DWARF type reader was updated to look for and use
the DW_AT_data_bit_offset attribute for structure fields, handling
the value of the attribute in the same way as for DW_AT_bit_offset.
This caused problems for clients, since the two attributes have very
different semantics.

This CL effectively reverts CL 328709 and moves to a scheme in which
we detect and report the two attributes separately.

Change-Id: Ic15ce01c851af38ebd81af827973ec49badcab6f