mime: add .ico to builtin types
Dear go maintainer.

I added the *.ico* extension as *image/x-icon" to the *builtinTypesLower* map on the *mime* package.

I think this fits as a trivial change and did not open an issue.

There's a discussion about this mime type: some say it is `image/vnd.microsoft.icon` but I've heard the most used is `image/x-icon`, so I decided to use the latter.

This is the first time I try to contribute to `go`, I hope this PR is correct.

Change-Id: I6a8bfc7cc8fa19adda1ccacfb5b52a9148ac713e
GitHub-Last-Rev: 139d86bddcb4cba59d3be2ed4138949155fceeba
GitHub-Pull-Request: golang/go#50802