runtime, sync, sync/atomic: document happens-before guarantees
A few of these are copied from the memory model doc.
Many are entirely new, following discussion on #47141.
See https://research.swtch.com/gomm for background.

The rule we are establishing is that each type that is meant
to help synchronize a Go program should document its
happens-before guarantees.

For #TBD.

DO NOT REVIEW - Comment on pending proposal #TBD.

Change-Id: I947c40639b263abe67499fa74f68711a97873a39