cmd/link/internal/loadpe: generalize handling of "__imp_*" syms
DO NOT SUBMIT

[not ready for review, needs testing]

The existing PE file loader has a special case for the function
__acrt_iob_func, whose hosting object file contains both an actual
definition and also a DLL import symbol __imp___acrt_iob_func. The
normal way of handling __imp_XXX symbols is the loader is to rename
them to their intended target (e.g. "XXX") however if the target is
also defined locally, you get a duplicate definition.

This patch generalizes the def/import symbol detection to apply to all
symbols in the object file being loaded (not just a hard-coded set),
since it will be needed when reading things like crt2.o.

Change-Id: I0d0607c27bb7d5f3cb415bc95db816aa13746ba2