cmd/link: read crt2.o for windows internal-linking CGO
DO NOT SUBMIT

[needs more testing, possibly some unit tests]
[needs testing on the windows 386 builder as well]

For Windows internal linking with CGO, when using more modern
LLVM-based compilers, read in the object file "crt2.o" so as to have a
definition of "atexit", needed to link the runtime/cgo test. The goal
here is to get all.bash working on Windows when using an up to date C
compiler; I am testing with this compiler, which is based on LLVM 13:

https://github.com/mstorsjo/llvm-mingw/releases/download/20211002/llvm-mingw-20211002-msvcrt-x86_64.zip

This patch also adds a new "hostObject" helper routine, similar to
"hostArchive" but specific to individual object files. There is also a
change to hostArchive to modify the pseudo-package name assigned when
reading archive elements: up until this point, a package name of
"libgcc" was used (even when reading a host archive like
"libmingex.a"), which led to very confusing errors messages if symbols
were missing or there were duplicate definitions.

Change-Id: I19c17dea9cfffa9e79030fc23064c7c63a612097