cmd/link/internal/loadpe: add rudimentary COMDAT support
DO NOT SUBMIT

[needs testing]

Add some rudimentary support to the PE file loader for handling
sections in COMDAT when reading host object files. This is needed
in order to link programs with support libraries that are of a more
modern vintage than GCC 5.X.

If a given section XYZ is in COMDAT, the symbol for that section will be
flagged (section Characteristics field has IMAGE_SCN_LNK_COMDAT bit set)
and the symbol will be followed by an "aux" symbol that includes (among
other things) the COMDAT handling strategy that the linker needs to use.
[Note: debug/pe does not support reading aux symbols, hence the code in
the linker has to use "unsafe" to dig them out at the moment].

This patch supports only a single strategy (IMAGE_COMDAT_SELECT_ANY); more
work will have to be done in the future to support other flavors if it
turns out that they are needed.

Change-Id: I516e825c30ed3df94ba08323b8a24fb847e10c1a