cmd/link: resolve __CTOR_LIST__/__DTOR_LIST__ refs from PE host objects
DO NOT SUBMIT

When doing an internal link on Windows, it's possible to see
unresolved references the symbols "__CTOR_LIST__" and/or
"__DTOR_LIST__" (which are needed in some circumstances). If these are
still unresolved at the point where we're done reading host objects,
then synthesize dummy versions of them.

Change-Id: I408bf18499bba05752710cf5a41621123bd84a3b